test: unit test sync pipeline
