+++ test/conftest.py
@@ -1,18 +1,37 @@
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-from pathlib import Path
-from typing import Generator
 
+from pathlib import Path
+from typing import Generator, Dict
+
+import asyncio
 import pytest
+import time
 
 from test.tracing.utils import SpyingTracer
 
 from haystack.testing.test_utils import set_all_seeds
-from haystack import tracing
+from haystack import tracing, component
 
 set_all_seeds(0)
 
+
+@pytest.fixture()
+def waiting_component():
+    @component
+    class Waiter:
+        @component.output_types(waited_for=int)
+        def run(self, wait_for: int) -> Dict[str, int]:
+            time.sleep(wait_for)
+            return {'waited_for': wait_for}
+
+        @component.output_types(waited_for=int)
+        async def run_async(self, wait_for: int) -> Dict[str, int]:
+            await asyncio.sleep(wait_for)
+            return {'waited_for': wait_for}
+
+    return Waiter
 
 @pytest.fixture()
 def test_files_path():
